refactor(backend): sales_quotations.customer_id BIGINT/i64 统一收窄为 INTEGER/i32
外键比被引用主键 customers.id(SERIAL i32)更宽是历史遗留(m0044 fix_fk_types 未覆盖 sales_crm),
虽 PG widening 兼容不崩,但属类型不一致+散落 as i32/as i64 强转。技术债清零:迁移两处 BIGINT→INTEGER、
模型/DTO/ListQuery 同步 i32、删 5 处强转。无 ALTER TYPE BIGINT 加宽覆盖(只命中 custom_orders,未碰)。
JSON 仍是数字,前端不破。cargo check --all-targets 0 错。

+++ backend/migration/src/domain/sales_crm/mod.rs
@@ -20,7 +20,7 @@ impl MigrationTrait for Migration {
 CREATE TABLE IF NOT EXISTS "sales_quotations" (
     "id" BIGSERIAL PRIMARY KEY,
     "quotation_no" VARCHAR(50) UNIQUE NOT NULL,
-    "customer_id" BIGINT NOT NULL REFERENCES "customers"("id"),
+    "customer_id" INTEGER NOT NULL REFERENCES "customers"("id"),
     "sales_user_id" BIGINT NOT NULL REFERENCES "users"("id"),
     "quotation_date" DATE NOT NULL,
     "valid_until" DATE NOT NULL,
