+++ backend/src/services/quotation_convert_service.rs
@@ -140,7 +140,7 @@ impl QuotationConvertService {
         let new_order = OrderActive {
             id: Default::default(),
             order_no: Set(order_no),
-            customer_id: Set(quotation.customer_id as i32),
+            customer_id: Set(quotation.customer_id),
             opportunity_id: Set(None),
             order_date: Set(now),
             required_date: Set(Utc::now() + chrono::Duration::days(30)),
